[Feat/#142] View와 ViewController 책임 분리
